requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.2
Requests==2.32.3
tabulate==0.9.0
